+++ common/src/main/java/net/mca/client/gui/VillagerEditorScreen.java
@@ -1910,7 +1910,11 @@ public class VillagerEditorScreen extends Screen implements SkinListUpdateListen
                 if (shouldPrintPlayerHint() && villagerUUID.equals(playerUUID) && getSelectedPlayerModel() != VillagerLike.PlayerModel.VILLAGER) {
                     final MatrixStack matrices = context.getMatrices();
                     matrices.push();
-                    matrices.translate(x, y - 127, 0);
+                    matrices.translate(
+                            (presetsButton.getX() + presetsButton.getWidth() + exportSkinButton.getX()) / 2.0F,
+                            presetsButton.getY() + presetsButton.getHeight() + 6,
+                            0
+                    );
                     matrices.scale(0.5f, 0.5f, 0.5f);
                     context.drawCenteredTextWithShadow(textRenderer, Text.translatable("gui.villager_editor.model_hint"), 0, 0, 0xAAFFFFFF);
                     matrices.pop();
@@ -2239,6 +2243,10 @@ public class VillagerEditorScreen extends Screen implements SkinListUpdateListen
     private NbtCompound saveEntityData(VillagerEntityMCA entity) {
         NbtCompound nbt = new NbtCompound();
         entity.writeCustomDataToNbt(nbt);
+        Text customName = entity.getCustomName();
+        if (customName != null) {
+            nbt.putString("CustomName", Text.Serializer.toJson(customName));
+        }
         return nbt;
     }